03-declarative-agent-manifest/endpoint/requirements.txt
flask>=3.0.0
gunicorn>=22.0.0
azure-ai-projects>=1.0.0b5
azure-identity>=1.15.0
PyJWT>=2.8.0
cryptography>=42.0.0
requests>=2.31.0
